requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
aiosqlite==0.19.0
stripe==7.8.0
asyncio==3.4.3
pytz==2023.3
apscheduler==3.10.4
